requirements.txt
py-cord==2.6.0
wavelink==2.6.3
python-dotenv==1.0.1
discord.py
dotenv